// src/site/site.dto.ts
export class SiteDto {
      title?: string;
      site_name?: string;
      body_background?: string;
      lead_name?: string;
      lead_name_color?: string;
      lead_subtitle?: string;
      lead_subtitle_color?: string;
      //lead_photo: string;
      name_color?: string;
      text_color?: string;
      about_name?: string;
      about_text?: string;
      //about_photo: string;
      client_name?: string;
      client_list?: string;
      photo_name?: string;
      plus_name?: string;
      plus_list?: string;
      plan_name?: string;
      plan_list?: string;
      button_name?: string;
      button_list?: string;
      contact_name?: string;
      contact_text?: string;
      phone_number?: string;
      vk?: string;
      tg?: string;
      mail?: string;
      address_name?: string;
      address?: string;
      map_link?: string;
    }
    
export class SiteDtoWithPath {
  title?: string;
  site_name?: string;
  icon_path?: string;
  body_background?: string;
  lead_name?: string;
  lead_name_color?: string;
  lead_subtitle?: string;
  lead_subtitle_color?: string;
  lead_photo_path?: string;
  name_color?: string;
  text_color?: string;
  about_name?: string;
  about_text?: string;
  about_photo_path: string;
  client_name?: string;
  client_list?: string;
  photo_name?: string;
  gallery_list_path?: string;
  plus_name?: string;
  plus_list?: string;
  plan_name?: string;
  plan_list?: string;
  button_name?: string;
  button_list?: string;
  contact_name?: string;
  contact_text?: string;
  phone_number?: string;
  vk?: string;
  tg?: string;
  mail?: string;
  address_name?: string;
  address?: string;
  map_link?: string;
}
// export class SiteDto {
//     title: string;
//     site_name: string;
//     body_background: string;
//     lead_name: string;
//     lead_name_color: string;
//     lead_subtitle: string;
//     lead_subtitle_color: string;
//     //lead_photo: string;
//     name_color: string;
//     text_color: string;
//     about_name: string;
//     about_text: string;
//     //about_photo: string;
//     client_name: string;
//     client_list: string;
//     photo_name: string;
//     plus_name: string;
//     plus_list: string;
//     plan_name: string;
//     plan_list: string;
//     button_name: string;
//     button_list: string;
//     contact_name: string;
//     contact_text: string;
//     phone_number: string;
//     vk: string;
//     tg: string;
//     mail: string;
//     address_name: string;
//     address: string;
//     map_link: string;
//   }